# legacy/e-masjid.my-2.0/api/cadangan-api/model/response.go
package model

type Response struct {
	Content interface{} `json:"content"`
	Total   int         `json:"total"`

	// TODO: add more fields
	// Empty            bool        `json:"empty"`
	// First            bool        `json:"first"`
	// Last             bool        `json:"last"`
	// Number           int         `json:"number"`
	// NumberOfElements int         `json:"numberOfElements"`
	// Pageable         struct {
	// 	PageNumber int `json:"pageNumber"`
	// 	PageSize   int `json:"pageSize"`
	// 	Sort       struct {
	// 		Empty    bool `json:"empty"`
	// 		Unsorted bool `json:"unsorted"`
	// 		Sorted   bool `json:"sorted"`
	// 	}
	// 	Offset int `json:"offset"`
	// }
	// Size int `json:"size"`
	// Sort struct {
	// 	Empty    bool `json:"empty"`
	// 	Unsorted bool `json:"unsorted"`
	// 	Sorted   bool `json:"sorted"`
	// }
	// TotalElements int `json:"totalElements"`
	// TotalPages    int `json:"totalPages"`
}